% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-transport.R
\name{apply_transport_weights}
\alias{apply_transport_weights}
\title{Apply Transportability Weights to NMA}
\usage{
apply_transport_weights(data, weights, config)
}
\arguments{
\item{data}{NMA data}

\item{weights}{Transportability weights}

\item{config}{Configuration}
}
\description{
Apply Transportability Weights to NMA
}
