% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-priors.R
\name{lfahfn_apply_prior}
\alias{lfahfn_apply_prior}
\title{Apply Informative Priors to Bayesian NMA}
\usage{
lfahfn_apply_prior(data, config)
}
\arguments{
\item{data}{NMA data}

\item{config}{Configuration}
}
\description{
Apply Informative Priors to Bayesian NMA
}
