% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-component.R
\name{lfahfn_component_model}
\alias{lfahfn_component_model}
\title{Run Component Network Meta-Analysis}
\usage{
lfahfn_component_model(nma_fit, sep = "+")
}
\arguments{
\item{nma_fit}{A standard NMA fit object from netmeta()}

\item{sep}{Separator used to denote combinations in treatment names}
}
\description{
Run Component Network Meta-Analysis
}
