% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-geometry.R
\name{lfahfn_contribution_matrix}
\alias{lfahfn_contribution_matrix}
\title{Calculate Network Contribution Matrix}
\usage{
lfahfn_contribution_matrix(nma_fit)
}
\arguments{
\item{nma_fit}{A standard NMA fit object from netmeta()}
}
\description{
Calculate Network Contribution Matrix
}
