% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-diagnostics.R
\name{lfahfn_diagnostics}
\alias{lfahfn_diagnostics}
\title{Run Comprehensive Diagnostics}
\usage{
lfahfn_diagnostics(nma_fit, data, config = setup_lfahfn())
}
\arguments{
\item{nma_fit}{Network meta-analysis object}

\item{data}{Original data}

\item{config}{Configuration}
}
\value{
List of diagnostic results
}
\description{
Run Comprehensive Diagnostics
}
