% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-priors.R
\name{lfahfn_empirical_prior}
\alias{lfahfn_empirical_prior}
\title{Generate Empirical Prior for Heterogeneity (Tau)}
\usage{
lfahfn_empirical_prior(
  outcome_type = c("objective", "semi-objective", "subjective"),
  intervention_type = c("pharma_vs_pharma", "pharma_vs_placebo", "non_pharma", "any")
)
}
\arguments{
\item{outcome_type}{Outcome type: "objective", "semi-objective", or "subjective"}

\item{intervention_type}{Comparison type: "pharma_vs_pharma", "pharma_vs_placebo", etc.}
}
\description{
Generate Empirical Prior for Heterogeneity (Tau)
}
