% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-plots.R
\name{lfahfn_forest}
\alias{lfahfn_forest}
\title{Forest Plot for LFAHFN}
\usage{
lfahfn_forest(x, ...)
}
\arguments{
\item{x}{LFAHFN object}

\item{...}{Additional arguments}
}
\description{
Forest Plot for LFAHFN
}
