% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-metareg.R
\name{lfahfn_metareg}
\alias{lfahfn_metareg}
\title{Run Network Meta-Regression}
\usage{
lfahfn_metareg(data, covariates = NULL, config = setup_lfahfn())
}
\arguments{
\item{data}{NMA data}

\item{covariates}{Character vector of covariate names}

\item{config}{Configuration object}
}
\description{
Run Network Meta-Regression
}
