% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-ml.R
\name{lfahfn_ml_covariate_selection}
\alias{lfahfn_ml_covariate_selection}
\title{Discover Non-Linear Covariates using Machine Learning}
\usage{
lfahfn_ml_covariate_selection(data, potential_covariates = NULL, n_trees = 500)
}
\arguments{
\item{data}{NMA dataset containing effect sizes (TE) and potential covariates}

\item{potential_covariates}{Vector of column names to evaluate}

\item{n_trees}{Number of trees in the forest}
}
\description{
Discover Non-Linear Covariates using Machine Learning
}
