% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-bayesian.R
\name{lfahfn_nodesplit_bayesian}
\alias{lfahfn_nodesplit_bayesian}
\title{Bayesian Node-Splitting Analysis}
\usage{
lfahfn_nodesplit_bayesian(data, config)
}
\arguments{
\item{data}{NMA data}

\item{config}{Configuration}
}
\description{
Bayesian Node-Splitting Analysis
}
