% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-ntsa.R
\name{lfahfn_ntsa}
\alias{lfahfn_ntsa}
\title{Network Trial Sequential Analysis (NTSA)}
\usage{
lfahfn_ntsa(nma_fit, alpha = 0.05, beta = 0.2, mcid = 0.2)
}
\arguments{
\item{nma_fit}{A standard NMA fit object}

\item{alpha}{Overall Type I error rate}

\item{beta}{Overall Type II error rate (default: 0.20 for 80\% power)}

\item{mcid}{Minimum Clinically Important Difference}
}
\description{
Network Trial Sequential Analysis (NTSA)
}
