% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-diagnostics.R
\name{lfahfn_publication_bias}
\alias{lfahfn_publication_bias}
\title{Publication Bias Assessment}
\usage{
lfahfn_publication_bias(nma_fit, data)
}
\arguments{
\item{nma_fit}{NMA object}

\item{data}{Data}
}
\value{
Publication bias results
}
\description{
Publication Bias Assessment
}
