% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-rob.R
\name{lfahfn_rob_flow}
\alias{lfahfn_rob_flow}
\title{Propagate Risk of Bias Across the Network}
\usage{
lfahfn_rob_flow(nma_fit, rob_data)
}
\arguments{
\item{nma_fit}{A standard NMA fit object from netmeta()}

\item{rob_data}{A data frame with columns `studlab` and `rob_score`}
}
\value{
A data frame containing the expected network risk of bias for each comparison
}
\description{
Propagate Risk of Bias Across the Network
}
