% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-robust.R
\name{lfahfn_robust_bayesian}
\alias{lfahfn_robust_bayesian}
\title{Robust Bayesian Network Meta-Analysis}
\usage{
lfahfn_robust_bayesian(data, config, df = 4)
}
\arguments{
\item{data}{NMA data}

\item{config}{Configuration}

\item{df}{Degrees of freedom for the t-distribution}
}
\description{
Robust Bayesian Network Meta-Analysis
}
