% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-plots.R
\name{lfahfn_transitivity_plot}
\alias{lfahfn_transitivity_plot}
\title{Transitivity Check Plot}
\usage{
lfahfn_transitivity_plot(data, covariate)
}
\arguments{
\item{data}{NMA data frame}

\item{covariate}{Name of the covariate}
}
\description{
Transitivity Check Plot
}
