% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-plots.R
\name{plot.lfahfn}
\alias{plot.lfahfn}
\title{Plot LFAHFN Results}
\usage{
\method{plot}{lfahfn}(x, type = c("forest", "network", "funnel", "rank", "transitivity"), ...)
}
\arguments{
\item{x}{LFAHFN object}

\item{type}{Type of plot}

\item{...}{Additional arguments}
}
\description{
Plot LFAHFN Results
}
