% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfahfn-core.R
\name{run_lfahfn_analysis}
\alias{run_lfahfn_analysis}
\title{Run LFAHFN Analysis}
\usage{
run_lfahfn_analysis(data, target_population = NULL, config = setup_lfahfn())
}
\arguments{
\item{data}{Data frame with NMA data}

\item{target_population}{Target population for transportability}

\item{config}{Configuration from setup_lfahfn()}
}
\value{
LFAHFN results object with class "lfahfn"
}
\description{
Run LFAHFN Analysis
}
